+++ spec/features/user/connect_social_account_spec.rb
@@ -1,0 +1,19 @@
+require "rails_helper"
+
+feature "Connect social account" do
+  context "oauth confirmed" do
+    context "user confirmed" do
+    end
+
+    context "user not confirmed" do
+    end
+  end
+
+  context "oauth not confirmed" do
+    context "user confirmed" do
+    end
+
+    context "user not confirmed" do
+    end
+  end
+end
