+++ spec/features/visitor/sign_up_with_social_spec_example.rb
